requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.0.0

# Optional type stubs
types-setuptools
pandas-stubs
